Refactored app startup dialogs into a combined component that manages when they show up to avoid conflicts
