+++ app/src/main/res/values/strings.xml
@@ -503,5 +503,5 @@
     <string name="database_backed_up">Database Backed up.</string>
     <string name="settings_screen_backup_and_restore">Backup and restore</string>
     <string name="donate">Donate</string>
-    <string name="donation_dialog_dismiss">No Thanks</string>
+    <string name="donation_dialog_dismiss">No, thanks</string>
 </resources>
